ARIES-975: semantic versioning tool serialisation id comparision is incorrect
git-svn-id: https://svn.apache.org/repos/asf/aries/trunk@1415277 13f79535-47bb-0310-9956-ffa450edef68

+++ versioning/versioning-checker/src/main/java/org/apache/aries/versioning/utils/ClassDeclaration.java
@@ -395,9 +395,9 @@ public class ClassDeclaration extends GenericDeclaration {
             // check to see whether the serializable id is the same
             //ignore if it is enum
             if ((!getAllSupers().contains(SemanticVersioningUtils.ENUM_CLASS) && (!old.getAllSupers().contains(SemanticVersioningUtils.ENUM_CLASS)))) {
-                Long oldValue = getSerialVersionUID(old);
-                Long curValue = getSerialVersionUID(this);
-                if ((oldValue.longValue() != curValue.longValue())) {
+                long oldValue = getSerialVersionUID(old);
+                long curValue = getSerialVersionUID(this);
+                if ((oldValue != curValue)) {
                     reasons.add("The serializable class is no longer back compatible as the value of SerialVersionUID has changed from " + oldValue + " to " + curValue + ".");
                 }
             }
